Add test for missing semicolon in import
PPI can misparse something like "use A use B;" as a "use A" with
arguments of "use" and "B".  If we trust PPI on this we will end up
tossing out "use B" when we process the includes. To avoid this we will
eval the include first. If an error is returned on the eval, we will
exclude this include. It can't be rewritten. It's possible something
wonky could happen because we haven't found two includes. We could also
just stop processing the file entirely in that case, but let's see how
far this gets us first.