add option to skip outputting empty import lists in use statements
This adds an option flag that, when enabled, causes perlimports to
stop outputting empty import lists for use statements where there is
no symbol to be imported.

While some projects and people want empty import lists to increase
clarity and prevent the used module from having default exports in the
future, others see them as unnecessary visual noise and would like to
avoid them.

The CLI option is named `--skip-empty-imports`, and the configuration
option is named `skip_empty_imports`. Both default to false,
maintaining the current behavior that provides more deterministic
clarity and future-proofs against export changes in the used modules.

Thanks to the existing code infrastructure, the implementation is
relatively straightforward. We just add new options and wire them up
so that they are available in the Include module when outputting.
Since there is already a dedicated code branch for modules that never
export or from which nothing is imported, we can modify the output
format string template there to differentiate between the respective
values of the new skip_empty_imports flag.

Add two tests to ensure that enabling and disabling explicitly work.
Reuse the existing original-imports.pl test data module, as it is a
fitting test asset.

Closes: #118
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>